+++ tt-train/sources/examples/python/transformers/data.py
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+
+# SPDX-License-Identifier: Apache-2.0
+
 import os
 import numpy as np
 from datasets import load_dataset
